Trading/GenX_FX_MASTER/requirements.txt
# GenX_FX Trading System Dependencies

# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3

# Database
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0

# Trading & Financial Data
ccxt==4.1.38
alpha-vantage==2.3.1
yfinance==0.2.22

# AI & Machine Learning
numpy==1.25.2
pandas==2.1.3
scikit-learn==1.3.2
tensorflow==2.14.0

# Data Processing
aiofiles==23.2.1
aiohttp==3.9.1
websockets==12.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
isort==5.12.0
flake8==6.1.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment
python-dotenv==1.0.0
pyyaml==6.0.1

# Monitoring & Logging
structlog==23.2.0